[REFACTOR] #6 - Response 클래스 모두 response package 로 이동
