+++ src/main/kotlin/com/example/smartshopping/domain/response/ReviewResponse.kt
@@ -1,0 +1,11 @@
+package com.example.smartshopping.domain.response
+
+import java.util.*
+
+data class ReviewResponse(
+    val nickName : String,
+    val productName : String,
+    val score : Int,
+    val reviewText : String,
+    val date : Date
+)
